format + change ids and names of fields

+++ database/migrations/2024_10_10_164658_create_apartments_table.php
@@ -16,7 +16,7 @@ return new class extends Migration {
             $table->string("type");
             $table->string("city");
             $table->string("address");
-            $table->integer("price_per_day");
+            $table->float("price_per_day");
 //            need to add table booking later
 //            $table->foreignId("booking_id");
             $table->timestamps();
